rewrite with new obkective function

+++ src/alchemy/yocto/alchemy/equilibria.hpp
@@ -32,9 +32,9 @@ namespace yocto
             vector<double>         C;      //!< local concentrations
             vector<double>         dC;     //!< delta concentration
             vector<bool>           active; //!< active species flag
-            vector<int>            beta;   //!< gradient for balancing, M
-            matrix<int>            W;      //!< Nu' * adjoint(Nu*Nu') * Nu, MxM
-            vector<int>            eta;    //!< integer descent direction for balancing, Nu2*beta
+            vector<double>         beta;   //!< gradient for balancing, M
+            matrix<double>         W;      //!< Nu' * adjoint(Nu*Nu') * Nu, MxM
+            vector<double>         eta;    //!< integer descent direction for balancing, Nu2*beta
             
             equilibrium & add( const string &name, const equilibrium_constant &K);
             equilibrium & add( const string &name, const double                K);
@@ -49,6 +49,8 @@ namespace yocto
             void normalize( array<double> &C0, const double t );
 
             friend std::ostream & operator<<( std::ostream &, const equilibria &);
+
+
 
 
         private:
